requirements.txt
streamlit
streamlit-chat
python-dotenv
google-generativeai
yfinance
pydantic
plotly
pandas
requests
protobuf
typing-extensions